requirements.txt
python-telegram-bot
tinyDB